// src/seed/interfaces/poke-response.interfaces.ts
// Generated by https://quicktype.io

export interface PokeResponse {
  count:    number;
  next:     string;
  previous: null;
  results:  Result[];
}

export interface Result {
  name: string;
  url:  string;
}

// Generated by https://quicktype.io

export interface IPokemons {
  _id:  string;
  name: string;
  no:   number;
  __v:  number;
}